resume-parser-service/requirements.txt
fastapi
uvicorn
google-cloud-storage
pymupdf
openai>=1.0.0
